requirements.txt
django-allauth==0.37.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-extensions==2.1.2
django-notifications-hq==1.5.0
django-tables2==2.0.3
django-webpack-loader
Django==2.1.5
Werkzeug==0.16.0
celery==4.3.0
redis==3.2.1
djangorestframework
